chanstate: group open channel kv tlv data
Rename the open channel TLV data helper into the KV open channel domain
and fold in the key locator TLV adapter. This keeps the open channel
storage records and their encode/decode helpers in one KV store file
without changing the exported helper API.
