--- chanstate/key_locator.go
@@ -1,57 +1,0 @@
-package chanstate
-
-import (
-	"io"
-
-	"github.com/lightningnetwork/lnd/keychain"
-	"github.com/lightningnetwork/lnd/tlv"
-)
-
-// keyLocRecord is a wrapper struct around keychain.KeyLocator to implement the
-// tlv.RecordProducer interface.
-type keyLocRecord struct {
-	keychain.KeyLocator
-}
-
-// Record creates a Record out of a KeyLocator using the passed Type and the
-// EKeyLocator and DKeyLocator functions. The size will always be 8 as
-// KeyFamily is uint32 and the Index is uint32.
-//
-// NOTE: This is part of the tlv.RecordProducer interface.
-func (k *keyLocRecord) Record() tlv.Record {
-	// Note that we set the type here as zero, as when used with a
-	// tlv.RecordT, the type param will be used as the type.
-	return tlv.MakeStaticRecord(
-		0, &k.KeyLocator, 8, EKeyLocator, DKeyLocator,
-	)
-}
-
-// EKeyLocator is an encoder for keychain.KeyLocator.
-func EKeyLocator(w io.Writer, val interface{}, buf *[8]byte) error {
-	if v, ok := val.(*keychain.KeyLocator); ok {
-		err := tlv.EUint32T(w, uint32(v.Family), buf)
-		if err != nil {
-			return err
-		}
-
-		return tlv.EUint32T(w, v.Index, buf)
-	}
-
-	return tlv.NewTypeForEncodingErr(val, "keychain.KeyLocator")
-}
-
-// DKeyLocator is a decoder for keychain.KeyLocator.
-func DKeyLocator(r io.Reader, val interface{}, buf *[8]byte, l uint64) error {
-	if v, ok := val.(*keychain.KeyLocator); ok {
-		var family uint32
-		err := tlv.DUint32(r, &family, buf, 4)
-		if err != nil {
-			return err
-		}
-		v.Family = keychain.KeyFamily(family)
-
-		return tlv.DUint32(r, &v.Index, buf, 4)
-	}
-
-	return tlv.NewTypeForDecodingErr(val, "keychain.KeyLocator", l, 8)
-}
